+++ src/app/job/page.tsx
@@ -1,9 +1,11 @@
+import BookmarkComponent from '@/features/job/bookmark-component';
 import JobPostingCard from '@/features/job/job-posting-card';
 const JobPage = () => {
   return (
     <div>
       JobPage
       <JobPostingCard />
+      <BookmarkComponent jobPostingId={1} />
     </div>
   );
 };
